Massive update for style checks and hints for decorators.
git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk/qooxdoo@15804 61aaf1ca-a10e-0410-b9fb-ac63ec25887f

+++ framework/source/class/qx/theme/modern/Decoration.js
@@ -674,7 +674,7 @@ qx.Theme.define("qx.theme.modern.Decoration",
 
     "progressive-table-header-cell" :
     {
-      decorator :  qx.ui.decoration.Double,
+      decorator :  qx.ui.decoration.Single,
 
       style :
       {
